requirements.txt
matplotlib>=2.2.2
six>=1.11.0
scipy>=1.1.0
h5py>=2.7.1
pandas>=0.22.0
numpy>=1.15.1
cryptography>=2.1.4
Pillow>=5.3.0
keras>=2.2.4
lmdb>=0.94
moviepy>=0.2.3.5
scikit_learn>=0.20.0
tensorflow-gpu>=1.10.0
opencv-python>=3.4.0.12
keras>=2.2.2
